Changing sqlconn to accept proper arguments for certain functions
sqlconn has three functions:
  sqlConnStmtBindParam
  sqlConnStmtStartBindParamTableValued
  sqlConnStmtBindCol
which accept a last argument of ssize_t* strlen, which is a pointer to the size of the data being inserted/updated (set by program in sqlConnStmtBindParam and sqlConnStmtStartBindParamTableValued) or the size of the data being retrieved (database sets value to pointer in sqlConnStmtBindCol).

These functons are basic wrappers around SQLBindParam and SQLBindCol, which want SQLINTEGER* instead of ssize_t*.

This is a problem because ssize_t is a typedef of int whereas SQLINTEGER is a typedef of long.

C doesn't complain but I'm refactoring sqlconn and was using C++ and the compiler complains (rightly so).

This probably wouldn't cause any problems because nothing we insert/select etc. would ever bigger than the max size of int, but there's zero reason to not fix this and the changes are simple (requiring SQLINTEGER* as a parameter instead of ssize_t*).